+++ demo/polmineR.R
@@ -1,5 +1,16 @@
 library(polmineR)
 
+watson.tts.process(
+  creds = "~/Lab/tmp/watson.json",
+  "What is the PolmIne Project about? We offer the code and the data we need to make progress. The polmIneR package offers: Counting, dispersion, kwic, concordances, co-occurrences.",
+  11
+)
+
+watson.tts.process(
+  creds = "~/Lab/tmp/watson.json",
+  "Ja das ist ja eine schöne Überraschung.",
+  10
+)
 
 # check whether CORPUS_REGISTRY environment variable is set
 Sys.getenv("CORPUS_REGISTRY")